Integrate VM into the sim. Fix: Added Call op to assembler Fix: Reworks to faulty VM logic Fix: Correct assembler color conversion Fix: Rewrote tests to work with new VM Removed: No more i32s in VM_Value Add: time_now, tick_now to imports.js for logging

+++ engine/vm_test.odin
@@ -22,7 +22,7 @@ vm_test_push_pop :: proc(t: ^testing.T) {
 	err := vm_run(&vm, program)
 	testing.expect_value(t, err, VM_Error.None)
 	v := sm.pop_back(&vm.stack)
-	testing.expect_value(t, v, i32(123))
+	testing.expect_value(t, v, f32(123))
 }
 
 
@@ -61,7 +61,7 @@ vm_test_add :: proc(t: ^testing.T) {
 	err := vm_run(&vm, program)
 	testing.expect_value(t, err, VM_Error.None)
 	v := sm.pop_back(&vm.stack)
-	testing.expect_value(t, v, i32(69))
+	testing.expect_value(t, v, f32(69))
 }
 
 @(test)
@@ -101,7 +101,7 @@ vm_test_sub_float :: proc(t: ^testing.T) {
 	err := vm_run(&vm, program)
 	testing.expect_value(t, err, VM_Error.None)
 	v := sm.pop_back(&vm.stack)
-	testing.expect_value(t, v, i32(1))
+	testing.expect_value(t, v, f32(1))
 }
 
 // TODO: In order to restore this test, A distinction must be made between jump and call
